Update managelayers.m
Significant efficiency gains in the cell-splitting section. Previously, a loop was used to split cells in dz, W, m, T, etc, which resized each of those variables many times. A significant portion of the GEMB processing time was spent on that loop, so I vectorized it and it's much faster now.