Update managelayers.m
Renaming indices X1 and X2 because X2 was repeatedly defined as X2=i, so there was no need for X2 at all. In this commit, X2 is replaced with i.

In this commit, X1 is renamed i_target as a more intuitive index name that implies this is where the mass will end up.